@@ -2,7 +2,8 @@ import app from 'flarum/admin/app';
 import Button from 'flarum/common/components/Button';
 import Dropdown from 'flarum/common/components/Dropdown';
 import ExtensionPage from 'flarum/admin/components/ExtensionPage';
-import icon from 'flarum/common/helpers/icon';
+import Icon from 'flarum/common/components/Icon';
+import Badge from 'flarum/common/components/Badge';
 import ItemList from 'flarum/common/utils/ItemList';
 
 export default class AuthSettingsPage extends ExtensionPage {
@@ -16,7 +17,7 @@ export default class AuthSettingsPage extends ExtensionPage {
     return [
       <div className="container">
         <div className="AuthSettingsPage">
-          <div className="Form">
+          <form className="Form">
             {this.buildSettingComponent({
               type: 'boolean',
               setting: 'fof-oauth.only_icons',
@@ -68,7 +69,7 @@ export default class AuthSettingsPage extends ExtensionPage {
             </div>
 
             {this.submitButton()}
-          </div>
+          </form>
         </div>
       </div>,
     ];
@@ -83,6 +84,9 @@ export default class AuthSettingsPage extends ExtensionPage {
       const showSettings = !!this.showing[name];
       const callbackUrl = `${app.forum.attribute('baseUrl')}/auth/${name}`;
 
+      const groupId = this.setting(`fof-oauth.${name}.group`)();
+      const selectedGroup = groupId ? app.store.getById('groups', groupId) : null;
+
       items.add(
         `fof-oauth.${name}`,
         <div className={`Provider ${enabled ? 'enabled' : 'disabled'} ${showSettings && 'showing'}`}>
@@ -92,26 +96,32 @@ export default class AuthSettingsPage extends ExtensionPage {
               setting: `fof-oauth.${name}`,
               label: (
                 <div>
-                  {icon(provider.icon)}
+                  <Icon name={provider.icon} />
                   <span>{app.translator.trans(`fof-oauth.lib.providers.${name}`)}</span>
                 </div>
               ),
             })}
 
-            {
-              <Button
-                className={`Button Button--rounded ${this.showing[name] && 'active'}`}
-                onclick={() => (this.showing[name] = !showSettings)}
-                aria-label={app.translator.trans('fof-oauth.admin.settings_accessibility_label', {
-                  name,
-                })}
-              >
-                {icon(`fas fa-cog`)}
-              </Button>
-            }
+            {enabled && selectedGroup && (
+              <div className="Provider--group">
+                {/*<Icon name={selectedGroup.icon() || 'fas fa-user-group'} />*/}
+                <Badge icon={selectedGroup.icon() || 'fas fa-user-group'} />
+                {selectedGroup.namePlural()}
+              </div>
+            )}
+
+            <Button
+              className={`Button Button--rounded ${this.showing[name] && 'active'}`}
+              onclick={() => (this.showing[name] = !showSettings)}
+              aria-label={app.translator.trans('fof-oauth.admin.settings_accessibility_label', {
+                name,
+              })}
+            >
+              <Icon name="fas fa-cog" />
+            </Button>
           </div>
 
-          <div className="Provider--settings">
+          <div className="Provider--settings" inert={!showSettings}>
             <div>
               <p>
                 {app.translator.trans(`fof-oauth.admin.settings.providers.${name}.description`, {
@@ -181,7 +191,7 @@ export default class AuthSettingsPage extends ExtensionPage {
             <Dropdown
               label={
                 selectedGroup
-                  ? [icon(selectedGroup.icon() || icons[selectedGroup.id()]), '\t', selectedGroup.namePlural()]
+                  ? [<Icon name={selectedGroup.icon() || icons[selectedGroup.id()]} />, '\t', selectedGroup.namePlural()]
                   : app.translator.trans('fof-oauth.admin.settings.providers.no_group_label')
               }
               buttonClassName="Button"
