@@ -1,6 +1,5 @@
 import app from 'flarum/forum/app';
 import Component, { ComponentAttrs } from 'flarum/common/Component';
-import icon from 'flarum/common/helpers/icon';
 import Button from 'flarum/common/components/Button';
 import Link from 'flarum/common/components/Link';
 import type Mithril from 'mithril';
@@ -8,6 +7,7 @@ import LinkedAccount from '../models/LinkedAccount';
 import User from 'flarum/common/models/User';
 import ProviderInfo from './ProviderInfo';
 import extractText from 'flarum/common/utils/extractText';
+import Icon from 'flarum/common/components/Icon';
 
 interface IAttrs extends ComponentAttrs {
   provider: LinkedAccount;
@@ -32,7 +32,7 @@ export default class LinkStatus extends Component<IAttrs> {
 
   view(vnode: Mithril.Vnode<IAttrs, this>): Mithril.Children {
     return (
-      <div className={`LinkedAccounts-Account LinkedAccounts-Account--${this.attrs.provider.name()}`}>
+      <div className={`LinkedAccountsList-item LinkedAccountsList-item--${this.attrs.provider.name()}`}>
         {this.iconView()}
         {this.statusView()}
         {this.actionView()}
@@ -43,7 +43,7 @@ export default class LinkStatus extends Component<IAttrs> {
   iconView() {
     return (
       <div className="LinkedAccountsList-item-icon">
-        {icon(this.attrs.provider.icon(), { className: `Provider-Icon Provider-Icon--${this.attrs.provider.name()}` })}
+        <Icon name={this.attrs.provider.icon()} className={`Provider-Icon Provider-Icon--${this.attrs.provider.name()}`} />
       </div>
     );
   }
