@@ -1,13 +1,13 @@
 import app from 'flarum/forum/app';
 import { extend } from 'flarum/common/extend';
-import UserSecurityPage from 'flarum/forum/components/UserSecurityPage';
+import type UserSecurityPage from 'flarum/forum/components/UserSecurityPage';
 import ItemList from 'flarum/common/utils/ItemList';
 import LinkedAccounts from '../components/LinkedAccounts';
 import type Mithril from 'mithril';
 
 export default function addLinkedAccountsToUserSecurityPage() {
-  extend(UserSecurityPage.prototype, 'settingsItems', function (items: ItemList<Mithril.Children>) {
-    if (this.user !== app.session.user && !app.forum.attribute('fofOauthModerate')) {
+  extend('flarum/forum/components/UserSecurityPage', 'settingsItems', function (this: UserSecurityPage, items: ItemList<Mithril.Children>) {
+    if (this.user?.id() !== app.session.user?.id() && !app.forum.attribute('fofOauthModerate')) {
       return;
     }
 
