@@ -4,9 +4,8 @@ import LogInButtons from 'flarum/forum/components/LogInButtons';
 import LogInButton from 'flarum/forum/components/LogInButton';
 import extractText from 'flarum/common/utils/extractText';
 import Tooltip from 'flarum/common/components/Tooltip';
-import LogInModal from 'flarum/forum/components/LogInModal';
-import SignUpModal from 'flarum/forum/components/SignUpModal';
-import ForumApplication from 'flarum/forum/ForumApplication';
+// import type LogInModal from 'flarum/forum/components/LogInModal';
+// import type SignUpModal from 'flarum/forum/components/SignUpModal';
 import { openOAuthPopup } from '../utils/popupUtils';
 
 export default function () {
@@ -59,7 +58,7 @@ export default function () {
     vdom.attrs.className += ' FoFLogInButtons--icons';
   });
 
-  extend(ForumApplication.prototype, 'authenticationComplete', function (_, payload) {
+  extend('flarum/forum/ForumApplication', 'authenticationComplete', function (_, payload) {
     if (payload.loggedIn) {
       app.fof_oauth_loginInProgress = true;
       // This will automatically be reset, as authenticationComplete also triggers a window reload.
@@ -68,7 +67,7 @@ export default function () {
     }
   });
 
-  ForumApplication.prototype.linkingComplete = async function () {
+  override('flarum/forum/ForumApplication', 'authenticationComplete', async function () {
     try {
       app.fof_oauth_linkingInProgress = true;
       m.redraw();
@@ -104,21 +103,21 @@ export default function () {
       app.fof_oauth_linkingInProgress = false;
       m.redraw();
     }
-  };
+  });
 
-  extend(LogInModal.prototype, 'onbeforeupdate', function (vnode) {
+  extend('flarum/forum/components/LogInModal', 'onbeforeupdate', function (vnode) {
     if (app.fof_oauth_loginInProgress) {
       this.loading = true;
     }
   });
 
-  extend(SignUpModal.prototype, 'onbeforeupdate', function (vnode) {
+  extend('flarum/forum/components/SignUpModal', 'onbeforeupdate', function (vnode) {
     if (app.fof_oauth_loginInProgress) {
       this.loading = true;
     }
   });
 
-  extend(SignUpModal.prototype, 'fields', function (items) {
+  extend('flarum/forum/components/SignUpModal', 'fields', function (items) {
     // If a suggested username was not provided by the OAuth service, display some help text to the user.
     if (!!this.attrs.token && !!!this.attrs.username) {
       items.add(
